Do you want magick?
I see a lot of notes like this when building the vignette. It's not an error, but just FYI in case your vignette would display better with magick. For example:

```
Summarizing rna molecular data for:	CCLE
The magick package is required to crop "/home/rstudio/vignettes/GxWorkshop_files/figure-html/ccleaac-1.png" but not available.
```